+++ jest.config.js
@@ -44,6 +44,10 @@ module.exports = {
     // in production builds. They are marked using the .skip-bundled.js extension.
     ...(isPublishBundle ? ["\\.skip-bundled\\.js$"] : []),
     ...(LIB_USE_ESM ? ["/babel-helpers/"] : []),
+    // Ignore @babel/standalone test in coverage testing because it is not built
+    ...(process.env.TEST_TYPE === "cov"
+      ? ["<rootDir>/packages/babel-standalone/"]
+      : []),
   ],
   testEnvironment: "node",
   transformIgnorePatterns: [
